Make it possible to run the corrode pipeline through the library
Moves all logic except a thin wrapper out of the executable Main module and into the library itself which makes it possible for the test module to run corrode directly instead of going through an external process. (Running corrode as an external process does not work at all from the tests since there is no way of knowing where the corrode executable is placed).